requirements.txt
Django==3.0.5
django-colorfield==0.5.0
djangorestframework==3.11.0
djoser==2.1.0
Pillow==8.4.0

